+++ data/jerry.txt
@@ -1,6 +1,4 @@
 D | 1 | quiz | 2022-10-10 12:30
 D | 0 | project | 2025-12-01 23:59
-D | 0 | report | 2025-10-10 09:00
-E | 0 | meeting | 2025-10-01 09:00 | 2025-10-01 10:00
-T | 0 | laundry
-D | 0 | assignment | 2022-10-10 10:30
+E | 1 | meeting | 2025-10-01 09:00 | 2025-10-01 10:00
+T | 0 | groceries
