+++ src/main/java/jerry/command/MarkCommand.java
@@ -1,16 +1,17 @@
-package command;
+package jerry.command;
 
-import exceptions.InvalidCommandFormatException;
-import exceptions.JerryException;
-import storage.Storage;
-import tasklist.TaskList;
-import ui.Ui;
+import jerry.exceptions.InvalidCommandFormatException;
+import jerry.exceptions.JerryException;
 
-public class UnmarkCommand extends Command {
+import jerry.storage.Storage;
+import jerry.tasklist.TaskList;
+import jerry.ui.Ui;
+
+public class MarkCommand extends Command {
 
     private int index = 0;
 
-    public UnmarkCommand(String input) throws InvalidCommandFormatException {
+    public MarkCommand(String input) throws InvalidCommandFormatException {
         String[] entries = input.split(" ", 2);
         if (entries.length < 2 || entries[1].trim().isEmpty()) {
             throw new InvalidCommandFormatException("Task number must be positive!");
@@ -24,7 +25,7 @@ public class UnmarkCommand extends Command {
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws JerryException {
-        this.response = taskList.unmark(index);
+        this.response = taskList.mark(index);
         taskList.saveTasks(storage);
         ui.displayOutput(this.response);
     }
