+++ src/main/java/jerry/command/UnmarkCommand.java
@@ -1,15 +1,16 @@
-package command;
+package jerry.command;
 
-import exceptions.InvalidCommandFormatException;
-import exceptions.JerryException;
-import storage.Storage;
-import tasklist.TaskList;
-import ui.Ui;
+import jerry.exceptions.InvalidCommandFormatException;
+import jerry.exceptions.JerryException;
+import jerry.storage.Storage;
+import jerry.tasklist.TaskList;
+import jerry.ui.Ui;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
+
     private int index = 0;
 
-    public DeleteCommand(String input) throws InvalidCommandFormatException {
+    public UnmarkCommand(String input) throws InvalidCommandFormatException {
         String[] entries = input.split(" ", 2);
         if (entries.length < 2 || entries[1].trim().isEmpty()) {
             throw new InvalidCommandFormatException("Task number must be positive!");
@@ -23,7 +24,7 @@ public class DeleteCommand extends Command {
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws JerryException {
-        this.response = taskList.deleteTask(index);
+        this.response = taskList.unmark(index);
         taskList.saveTasks(storage);
         ui.displayOutput(this.response);
     }
