+++ src/main/java/jerry/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package jerry.exceptions;
 
 public class InvalidCommandException extends JerryException {
     public InvalidCommandException(String message) {
