+++ src/main/java/jerry/exceptions/JerryException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package jerry.exceptions;
 
 public class JerryException extends Exception {
     public JerryException(String message) {
