+++ src/main/java/jerry/exceptions/ListIndexOutOfBoundException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package jerry.exceptions;
 
 public class ListIndexOutOfBoundException extends JerryException {
     public ListIndexOutOfBoundException(String message) {
