+++ src/main/java/jerry/task/Event.java
@@ -1,6 +1,6 @@
-package task;
+package jerry.task;
 
-import exceptions.InvalidCommandFormatException;
+import jerry.exceptions.InvalidCommandFormatException;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
